build: disable validation
Documentation:
https://wixtoolset.org/documentation/manual/v3/overview/light.html

This may be a limitation of the validation system when the build is ran
as a service (buildbot-worker) without Admin privilege:
https://github.com/wixtoolset/issues/issues/3968

Since running smoke.exe
(https://wixtoolset.org/documentation/manual/v3/overview/alltools.html)
in a command console, using the same de-privileged user, succeeds, this
is likely something wrong in the installer itself (install.wxs).

Turn off the validation until this is sorted.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>